Adds in diffHTML example
diffHTML is a React-inspired tiny library that efficiently diffs HTML
to the DOM and patches in changes. It also provides a Custom Element
polyfill and a transitions API.

While it's still under development, it has already been successfully
used on an internal Netflix application.

Adding to your list in hopes that I'll encourage others who are
interested to contribute.

Relevant links:

[http://github.com/tbranyen/diffhtml](http://github.com/tbranyen/diffhtml)
[http://diffhtml.org](http://diffhtml.org)